// src/api/relation/relation.message.ts
export enum RelationMessage {
  NOT_FOUND_USER = '유저를 찾을 수 없습니다.',
  NOT_FOUND_RELATION = '관계 정보를 찾을 수 없습니다.',
  NOT_FOLLOW_RELATION = 'FOLLOW 상태가 아닙니다.',
  NOT_BLOCKED_RELATION = '차단 상태가 아닙니다.',
  FOLLOW_RELATION = '이미 FOLLOW 상태입니다.',
  UNAUTHORIZATION_RELATION = 'FOLLOW 할 수 있는 상태가 아닙니다.',
}